chore: moved api resource to UI namespace
