+++ config/packages/api_platform.yaml
@@ -1,0 +1,4 @@
+api_platform:
+    mapping:
+        paths:
+            - '%kernel.project_dir%/src/UI/ApiResource'
